+++ playground/textarea.php
@@ -4,11 +4,11 @@ use function Laravel\Prompts\textarea;
 
 require __DIR__.'/../vendor/autoload.php';
 
-$email = textarea(
+$story = textarea(
     label: 'Tell me a story',
     placeholder: 'Weave me a tale',
 );
 
-var_dump($email);
+var_dump($story);
 
 echo str_repeat(PHP_EOL, 5);