+++ src/Concerns/TypedValue.php
@@ -27,8 +27,10 @@ trait TypedValue
             $this->cursorPosition = mb_strlen($this->typedValue);
         }
 
-        $this->on('key', function ($key) use ($submit, $ignore, $allowNewLine) {
-            if ($key[0] === "\e" || in_array($key, [Key::CTRL_B, Key::CTRL_F, Key::CTRL_A, Key::CTRL_E])) {
+        $this->on('key', function (string $key) use ($submit, $ignore, $allowNewLine): void {
+            if ($key !== '' &&
+                ($key[0] === "\e" || in_array($key, [Key::CTRL_B, Key::CTRL_F, Key::CTRL_A, Key::CTRL_E]))
+            ) {
                 if ($ignore !== null && $ignore($key)) {
                     return;
                 }